Move code from toot.tui.base to toot.tui
