+++ toot/cli/accounts.py
@@ -4,7 +4,7 @@ import json as pyjson
 from typing import BinaryIO, Optional
 
 from toot import api
-from toot.cli.base import PRIVACY_CHOICES, cli, json_option, Context, pass_context
+from toot.cli import PRIVACY_CHOICES, cli, json_option, Context, pass_context
 from toot.cli.validators import validate_language
 from toot.output import print_acct_list
